Standardized ability score abbreviations to be sentence case
When the ability score abbreviation is used on its own or as a +Score
there was variation in displaying it. In most cases, the score was in
sentence case (e.g. Str) more often than it was in all caps (e.g. STR).
I made the less-frequent all caps version into sentence case.